backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-generativeai==0.7.2
langchain==0.2.6
langgraph==0.1.13
langsmith==0.1.77
chromadb==0.5.3
guardrails-ai==0.5.6
pydantic==2.7.4
aiohttp==3.9.5
python-dotenv==1.0.1
faker==25.2.0
